requirements.txt
azure-cognitiveservices-language-textanalytics==0.2.0
azure-common==1.1.23
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.2.4
djangorestframework==3.10.2
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
msrest==0.6.9
numpy==1.17.0
oauthlib==3.1.0
opencv-python==3.4.5.20
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
tqdm==4.34.0
urllib3==1.25.3
watson-developer-cloud==2.10.1
websocket-client==0.48.0
Werkzeug==0.15.5
